news_recommendation/requirements.txt
asgiref==3.2.7
daphne==2.5.0
Django==3.0.6
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
torch==1.5.0
torchvision==0.6.0 